Fixed the parsing 'string' tokens. There were a couple of problems causing strings like '\\' and '\'' to be parsed incorrectly.
First, in case a quote char was found, there was first a check to see if it was followed by the same quote char, and if it did it was assumed that that formed an escape sequence. This is wrong, because there was no check to see if the first quote might have been a literal quote due to a preceding slash.

If no subsequent quote char would be found, the code would check if the quote was preceded by a slash, assuming that quote would be a literal quote escaped by that slash. However, this assumption could be wrong, since that slash might itself be escaped by a preceding slash, making it a literal slash.

I think I fixed it though:
- when we see a quote, use locate to find the next quote. 
- if we don't find it, we obviously have an unterminated string, so we bail out in the error state.
- if we do find a quote, we must first check if that quote was escaped by a preceding backslash. This must be done with a loop: if we find a sequence of an odd number of backslashes immediately before the quote, then the quote is a literal quote that was escaped by its preceding backslash and we should continue to scan for the real closing quote. 
- if the quote was not escaped by a preceding backslash, it can still be immediately followed by a quote char. If that is the case, we need to continue scanning for the real closing quote.
- if we find that we must continue to scan beyond the found quote, we must check if we're still in range, because if the string happens to end immediately after the escaped quote, the string is not terminated and we must deliver the error token.


git-svn-id: https://common-schema.googlecode.com/svn/trunk@497 c8a82860-ae11-f3b5-05a3-c0c726401f9a